created the post for posts
